Fix: Corrigindo namespace de todas as requests e importando corretamente nas devidas controllers
