+++ app/Http/Requests/Albums/StoreAlbumsRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Albums;
 
 use Illuminate\Foundation\Http\FormRequest;
