+++ app/Http/Requests/Fabricante/UpdateFabricanteRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Fabricante;
 
 use Illuminate\Foundation\Http\FormRequest;
