+++ app/Http/Requests/FileUser/UpdateFileUserRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\FileUser;
 
 use Illuminate\Foundation\Http\FormRequest;
