+++ app/Http/Requests/Parceiros/StoreParceirosRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Parceiros;
 
 use Illuminate\Foundation\Http\FormRequest;
