+++ app/Http/Requests/Parceiros/UpdateParceirosRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Parceiros;
 
 use Illuminate\Foundation\Http\FormRequest;
