+++ app/Http/Requests/PerfilAcessos/StorePerfilAcessosRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\PerfilAcessos;
 
 use Illuminate\Foundation\Http\FormRequest;
