+++ app/Http/Requests/Permissoes/UpdatePermissoesRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Permissoes;
 
 use Illuminate\Foundation\Http\FormRequest;
