+++ app/Http/Requests/Publicacao/UpdatePublicacaosRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Publicacao;
 
 use Illuminate\Foundation\Http\FormRequest;
