+++ app/Http/Requests/RedeSocial/StoreRedeSocialRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\RedeSocial;
 
 use Illuminate\Foundation\Http\FormRequest;
