+++ app/Http/Requests/RegistroExecTreino/StoreRegistroExecTreinoRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\RegistroExecTreino;
 
 use Illuminate\Foundation\Http\FormRequest;
