+++ app/Http/Requests/RegistroExecTreino/UpdateRegistroExecTreinoRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\RegistroExecTreino;
 
 use Illuminate\Foundation\Http\FormRequest;
