+++ app/Http/Requests/TagPublicacao/UpdateTagPublicacaoRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\TagPublicacao;
 
 use Illuminate\Foundation\Http\FormRequest;
